Move the free port logic so that loadOptions don't override it (#7237)
Move the free port logic so that loadOptions don't override it
