+++ lib/core/src/server/build-dev.js
@@ -16,6 +16,8 @@ import semver from 'semver';
 import { stripIndents } from 'common-tags';
 import Table from 'cli-table3';
 import prettyTime from 'pretty-hrtime';
+import inquirer from 'inquirer';
+import detectFreePort from 'detect-port';
 
 import storybook from './dev-server';
 import { getDevCli } from './cli';
@@ -34,6 +36,12 @@ const writeStats = async (name, stats) => {
     'utf8'
   );
 };
+
+const getFreePort = port =>
+  detectFreePort(port).catch(error => {
+    logger.error(error);
+    process.exit(-1);
+  });
 
 async function getServer(app, options) {
   if (!options.https) {
@@ -238,7 +246,24 @@ function openInBrowser(address) {
 
 export async function buildDevStandalone(options) {
   try {
-    const { port, host, extendServer } = options;
+    const { host, extendServer } = options;
+
+    const port = await getFreePort(options.port);
+
+    if (!options.ci && !options.smokeTest && options.port != null && port !== options.port) {
+      const { shouldChangePort } = await inquirer.prompt({
+        type: 'confirm',
+        default: true,
+        name: 'shouldChangePort',
+        message: `Port ${
+          options.port
+        } is not available. Would you like to run Storybook on port ${port} instead?`,
+      });
+
+      if (!shouldChangePort) {
+        process.exit(1);
+      }
+    }
 
     // Used with `app.listen` below
     const listenAddr = [port];
