+++ lib/core/src/server/cli/dev.js
@@ -1,15 +1,8 @@
 import program from 'commander';
 import chalk from 'chalk';
-import detectFreePort from 'detect-port';
 import inquirer from 'inquirer';
 import { logger } from '@storybook/node-logger';
 import { parseList, getEnvConfig } from './utils';
-
-const getFreePort = port =>
-  detectFreePort(port).catch(error => {
-    logger.error(error);
-    process.exit(-1);
-  });
 
 async function getCLI(packageJson) {
   process.env.NODE_ENV = process.env.NODE_ENV || 'development';
@@ -62,22 +55,7 @@ async function getCLI(packageJson) {
     program.port = parseInt(program.port, 10);
   }
 
-  const port = await getFreePort(program.port);
-
-  if (!program.ci && !program.smokeTest && program.port != null && port !== program.port) {
-    const { shouldChangePort } = await inquirer.prompt({
-      type: 'confirm',
-      default: true,
-      name: 'shouldChangePort',
-      message: `Port ${program.port} is not available.
-Would you like to run Storybook on port ${port} instead?`,
-    });
-    if (!shouldChangePort) {
-      process.exit(1);
-    }
-  }
-
-  return { ...program, port };
+  return { ...program };
 }
 
 export default getCLI;
